requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
celery-progress==0.5
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.27.0.post1  # https://github.com/encode/uvicorn
pyyaml==6.0.1
channels==4.0.0
channels-redis==4.2.0
ping3==4.0.4

# TODO : remove that old version
# Django
# ------------------------------------------------------------------------------
django==4.2.10  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2023.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-widget-tweaks==1.5.0
# Django
# ------------------------------------------------------------------------------
#django==5.1.5 #==5.0.11  # pyup: < 5.1  # https://www.djangoproject.com/
#django-environ==0.12.0  # https://github.com/joke2k/django-environ
#django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
#django-allauth[mfa]==65.3.1  # https://github.com/pennersr/django-allauth
#django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
#crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
#django-redis==5.4.0  # https://github.com/jazzband/django-redis
#django-widget-tweaks==1.5.0
